Switch examples to H.264 MP4
This miraculously makes the Reverse example work for me on macOS (the previous example file was a Theora video in a MKV container). Using the new example file I am moreover unable to make the Scratch example stop giving us frames, which it did previously (H.264 in a QuickTime MOV).

Since H.264 in a MP4 container seems to be the most reasonable default these days anyhow, let's try at least for a while. Feel free to re-compress the example file to make it smaller though.